feat(good-tracing): add basic batch control tab for good consumption
Adds a batch control tab that shows all buildings consuming goods, grouped by good.
The rationale is to include buildings that have the potential of consuming a certain good.
Manufacturies are included only if the current recipe is using the good.
Buildings are included even if they are not actively consuming the good (e.g. because
they are paused, lack workers or power, etc..).
Stockpiles are excluded.