feat(good-tracing): add basic batch control tab for good production
Adds a batch control tab showing all buildings that produce goods, grouped by good.
The rationale is to display buildings that have the potential of producing a certain good.
Manufacturies are showed only if the current recipe is producing the good.
Buildings are shown even if they are not actively producing (e.g. because they are paused,
lack workers/power, etc..).
Gatherer flags, farmhouses and similar show up under a given good only if the good is actually
in range to be gathered.
Stockpiles and district crossings are excluded.