refactor(good-tracing): improved management of "show paused buildings"
Paused buildings are always added to the row groups and their visibility
is now controlled via the VisibilityGetter property of BatchControlRow.
This avoids refreshing the whole UI when the "show paused buildings" checkbox
is ticked/unticked, and also allows to refresh the UI in real time as buildings
paused state changes.