fix(good-tracing): fix issues caused by buildings transitioning to finished state
Unfinished buildings that still require some goods are always added to the respective row groups.
Rows are refreshed when a building transitions to finished state.
Visibility check reconsiders whether a building should have been added to any row group
(this is necessary because it may be possible for a building to be added because it was under construction,
but normally you would not add it, so in this case just don't display it).