feat(construction-queue): base implementation of the UI
Show the list of construction sites on the top-left.
Rows are setup as batch control rows.
Selected construction sites are highlighted.
The statuses mostly update themselves correctly, except for a few.
Needs to be optimized (right now I clear and rebuild the list every time,
which also clears the class for the selected row).